Pull parts related code into s.i.parts
Start refactoring parts loading logic by separating it into
snapcraft.internal.parts.

LP:#1596044
