+++ snapcraft/internal/parts.py
@@ -23,8 +23,11 @@ import requests
 import yaml
 from xdg import BaseDirectory
 
+import snapcraft
 from snapcraft.internal.indicators import download_requests_stream
 from snapcraft.internal.common import get_terminal_width
+from snapcraft.internal import sources, pluginhandler
+from snapcraft.internal import yaml as internalyaml
 
 
 PARTS_URI = 'https://parts.snapcraft.io/v1/parts.yaml'
@@ -127,6 +130,206 @@ class _RemoteParts(_Base):
         remote_part.update(properties)
 
         return remote_part
+
+
+class SnapcraftLogicError(Exception):
+
+    @property
+    def message(self):
+        return self._message
+
+    def __init__(self, message):
+        self._message = message
+
+
+def _expand_filesets_for(step, properties):
+    filesets = properties.get('filesets', {})
+    fileset_for_step = properties.get(step, {})
+    new_step_set = []
+
+    for item in fileset_for_step:
+        if item.startswith('$'):
+            try:
+                new_step_set.extend(filesets[item[1:]])
+            except KeyError:
+                raise SnapcraftLogicError(
+                    '\'{}\' referred to in the \'{}\' fileset but it is not '
+                    'in filesets'.format(item, step))
+        else:
+            new_step_set.append(item)
+
+    return new_step_set
+
+
+class PartsConfig:
+
+    def __init__(self, parts_data, project_options, validator, build_tools,
+                 snapcraft_yaml):
+        if project_options is None:
+            project_options = snapcraft.ProjectOptions()
+
+        self._parts_data = parts_data
+        self._project_options = project_options
+        self._validator = validator
+        self.build_tools = build_tools
+        self._snapcraft_yaml = snapcraft_yaml
+
+        self.all_parts = []
+        self._part_names = []
+        self.after_requests = {}
+
+        self._remote_parts = get_remote_parts()
+        self._process_parts()
+
+    def _process_parts(self):
+        for part_name in self._parts_data:
+            self._part_names.append(part_name)
+            properties = self._parts_data[part_name] or {}
+
+            plugin_name = properties.pop('plugin', None)
+            if not plugin_name:
+                logger.info(
+                    'Searching in the remote parts cache for part '
+                    '{!r}'.format(part_name))
+                try:
+                    properties = self._remote_parts.compose(
+                        part_name, properties)
+                    plugin_name = properties.pop('plugin', None)
+                except KeyError as e:
+                    raise SnapcraftLogicError(
+                        '{!r} is missing the `plugin` entry and is not '
+                        'defined in the current remote parts cache, try to '
+                        'run `snapcraft update` to '
+                        'refresh'.format(part_name)) from e
+
+            if 'after' in properties:
+                self.after_requests[part_name] = properties.pop('after')
+
+            properties['stage'] = _expand_filesets_for('stage', properties)
+            properties['snap'] = _expand_filesets_for('snap', properties)
+
+            if 'filesets' in properties:
+                del properties['filesets']
+
+            self.load_plugin(part_name, plugin_name, properties)
+
+        self._compute_part_dependencies()
+        self.all_parts = self._sort_parts()
+
+    def _compute_part_dependencies(self):
+        '''Gather the lists of dependencies and adds to all_parts.'''
+
+        for part in self.all_parts:
+            dep_names = self.after_requests.get(part.name, [])
+            for dep in dep_names:
+                found = False
+                for i in range(len(self.all_parts)):
+                    if dep == self.all_parts[i].name:
+                        part.deps.append(self.all_parts[i])
+                        found = True
+                        break
+                if not found:
+                    try:
+                        remote_part = self._remote_parts.get_part(dep)
+                    except KeyError as e:
+                        raise SnapcraftLogicError(
+                            'Cannot find definition for part {!r}. '
+                            'It may be a remote part, run `snapcraft update` '
+                            'to refresh the remote parts '
+                            'cache'.format(dep)) from e
+                    plugin_name = remote_part.pop('plugin')
+                    part.deps.append(self.load_plugin(
+                        dep, plugin_name, remote_part))
+                    self._part_names.append(dep)
+
+    def _sort_parts(self):
+        '''Performs an inneficient but easy to follow sorting of parts.'''
+        sorted_parts = []
+
+        while self.all_parts:
+            top_part = None
+            for part in self.all_parts:
+                mentioned = False
+                for other in self.all_parts:
+                    if part in other.deps:
+                        mentioned = True
+                        break
+                if not mentioned:
+                    top_part = part
+                    break
+            if not top_part:
+                raise SnapcraftLogicError(
+                    'circular dependency chain found in parts definition')
+            sorted_parts = [top_part] + sorted_parts
+            self.all_parts.remove(top_part)
+
+        return sorted_parts
+
+    def part_prereqs(self, part_name):
+        """Returns a set with all of part_names' prerequisites."""
+        return set(self.after_requests.get(part_name, []))
+
+    def part_dependents(self, part_name):
+        """Returns a set of all the parts that depend upon part_name."""
+
+        dependents = set()
+        for part, prerequisites in self.after_requests.items():
+            if part_name in prerequisites:
+                dependents.add(part)
+
+        return dependents
+
+    def get_part(self, part_name):
+        for part in self.all_parts:
+            if part.name == part_name:
+                return part
+
+        return None
+
+    def validate_parts(self, part_names):
+        for part_name in part_names:
+            if part_name not in self._part_names:
+                raise EnvironmentError(
+                    'The part named {!r} is not defined in '
+                    '{!r}'.format(part_name, self._snapcraft_yaml))
+
+    def load_plugin(self, part_name, plugin_name, properties):
+        part = pluginhandler.load_plugin(
+            part_name, plugin_name, properties,
+            self._project_options, self._validator.part_schema)
+
+        self.build_tools += part.code.build_packages
+        self.build_tools += sources.get_required_packages(part.code.options)
+        self.all_parts.append(part)
+        return part
+
+    def build_env_for_part(self, part, root_part=True):
+        """Return a build env of all the part's dependencies."""
+
+        env = []
+        stagedir = self._project_options.stage_dir
+
+        if root_part:
+            # this has to come before any {}/usr/bin
+            env += part.env(part.installdir)
+            env += internalyaml._runtime_env(
+                part.installdir, self._project_options.arch_triplet)
+            env += internalyaml._runtime_env(
+                stagedir, self._project_options.arch_triplet)
+            env += internalyaml._build_env(
+                part.installdir, self._project_options.arch_triplet)
+            env += internalyaml._build_env_for_stage(
+                stagedir, self._project_options.arch_triplet)
+        else:
+            env += part.env(stagedir)
+            env += internalyaml._runtime_env(
+                stagedir, self._project_options.arch_triplet)
+
+        for dep_part in part.deps:
+            env += dep_part.env(stagedir)
+            env += self.build_env_for_part(dep_part, root_part=False)
+
+        return env
 
 
 def update():
